+++ src/bin/development.rs
@@ -1,86 +1,93 @@
 //! Developement tests
 //!
 
-extern crate openslide;
 extern crate failure;
 extern crate image;
-#[macro_use] extern crate clap;
+extern crate openslide;
+#[macro_use]
+extern crate clap;
 
-use std::path::Path;
 use std::fs;
+use std::path::Path;
 
-use failure::{Error, format_err};
-use clap::{Arg, App, ArgMatches};
+use clap::{App, Arg, ArgMatches};
+use failure::{format_err, Error};
 use openslide::OpenSlide;
-
 
 fn get_cli<'a>() -> ArgMatches<'a> {
     let matches = App::new("Analyse wsi")
-                          .version("0.2.0")
-                          .author("Ole-Johan Skrede")
-                          .about("Toy program used for development of openslide.")
-                          .arg(Arg::with_name("input_file")
-                               .short("i")
-                               .long("input_file")
-                               .value_name("FILE")
-                               .help("Filename of input wsi image")
-                               .required(true)
-                               )
-                          .arg(Arg::with_name("out_root_dir")
-                               .short("o")
-                               .long("out_root_dir")
-                               .value_name("FOLDER")
-                               .default_value("/tmp/analyse_wsi")
-                               .help("Root directory of output images")
-                               )
-                          .arg(Arg::with_name("source_row")
-                               .short("r")
-                               .long("source_row")
-                               .value_name("UINT32")
-                               .default_value("0")
-                               .help("Crop start row in target pixels")
-                               )
-                          .arg(Arg::with_name("source_column")
-                               .short("c")
-                               .long("source_column")
-                               .value_name("UINT32")
-                               .default_value("0")
-                               .help("Crop end row in target pixels")
-                               )
-                          .arg(Arg::with_name("target_height")
-                               .short("h")
-                               .long("target_height")
-                               .allow_hyphen_values(true)
-                               .value_name("INT32")
-                               .default_value("512")
-                               .help("Size of crop in target pixels. -1 indicates full height")
-                               )
-                          .arg(Arg::with_name("target_width")
-                               .short("w")
-                               .long("target_width")
-                               .allow_hyphen_values(true)
-                               .value_name("INT32")
-                               .default_value("512")
-                               .help("Size of crop in target pixels. -1 indicates full width")
-                               )
-                          .arg(Arg::with_name("zoom_factor")
-                               .short("z")
-                               .long("zoom_factor")
-                               .value_name("FLOAT32")
-                               .default_value("1.0")
-                               .help("Zoom factor")
-                               )
-                          .arg(Arg::with_name("print_properties")
-                               .short("p")
-                               .long("print_properties")
-                               .takes_value(false)
-                               .help("Whether or not to print the slide properties")
-                               )
-                          .get_matches();
+        .version("0.2.0")
+        .author("Ole-Johan Skrede")
+        .about("Toy program used for development of openslide.")
+        .arg(
+            Arg::with_name("input_file")
+                .short("i")
+                .long("input_file")
+                .value_name("FILE")
+                .help("Filename of input wsi image")
+                .required(true),
+        )
+        .arg(
+            Arg::with_name("out_root_dir")
+                .short("o")
+                .long("out_root_dir")
+                .value_name("FOLDER")
+                .default_value("/tmp/analyse_wsi")
+                .help("Root directory of output images"),
+        )
+        .arg(
+            Arg::with_name("source_row")
+                .short("r")
+                .long("source_row")
+                .value_name("UINT32")
+                .default_value("0")
+                .help("Crop start row in target pixels"),
+        )
+        .arg(
+            Arg::with_name("source_column")
+                .short("c")
+                .long("source_column")
+                .value_name("UINT32")
+                .default_value("0")
+                .help("Crop end row in target pixels"),
+        )
+        .arg(
+            Arg::with_name("target_height")
+                .short("h")
+                .long("target_height")
+                .allow_hyphen_values(true)
+                .value_name("INT32")
+                .default_value("512")
+                .help("Size of crop in target pixels. -1 indicates full height"),
+        )
+        .arg(
+            Arg::with_name("target_width")
+                .short("w")
+                .long("target_width")
+                .allow_hyphen_values(true)
+                .value_name("INT32")
+                .default_value("512")
+                .help("Size of crop in target pixels. -1 indicates full width"),
+        )
+        .arg(
+            Arg::with_name("zoom_factor")
+                .short("z")
+                .long("zoom_factor")
+                .value_name("FLOAT32")
+                .default_value("1.0")
+                .help("Zoom factor"),
+        )
+        .arg(
+            Arg::with_name("print_properties")
+                .short("p")
+                .long("print_properties")
+                .takes_value(false)
+                .help("Whether or not to print the slide properties"),
+        )
+        .get_matches();
 
     matches
 }
-
 
 fn write_region(
     fname: &str,
@@ -92,12 +99,22 @@ fn write_region(
     target_width: i32,
     zoom_factor: f32,
 ) -> Result<(), Error> {
-
     let zoom_lvl = os.get_best_level_for_downsample(zoom_factor as f64)?;
     println!("Max number of levels: {}", os.get_level_count()?);
-    println!("Best zoom level for zoom factor {} is: {}", zoom_factor, zoom_lvl);
-    println!("Level {} downsample: {}", zoom_lvl, os.get_level_downsample(zoom_lvl as u8)?);
-    println!("Level {} dimensions: {:?}", zoom_lvl, os.get_level_dimensions(zoom_lvl)?);
+    println!(
+        "Best zoom level for zoom factor {} is: {}",
+        zoom_factor, zoom_lvl
+    );
+    println!(
+        "Level {} downsample: {}",
+        zoom_lvl,
+        os.get_level_downsample(zoom_lvl as u8)?
+    );
+    println!(
+        "Level {} dimensions: {:?}",
+        zoom_lvl,
+        os.get_level_dimensions(zoom_lvl)?
+    );
 
     let target_height = if target_height == -1 {
         let (_, full_height) = os.get_level_dimensions(zoom_lvl)?;
@@ -115,7 +132,13 @@ fn write_region(
 
     println!("Read region with height: {}", target_height);
     println!("Read region with width: {}", target_width);
-    let im = os.read_region(source_row, source_col, zoom_lvl, target_height, target_width)?;
+    let im = os.read_region(
+        source_row,
+        source_col,
+        zoom_lvl,
+        target_height,
+        target_width,
+    )?;
     /*
     let out_fname = out_dir.join(format!("wsi_region_x{}_y{}_h{}_w{}_z{}.png",
                                  source_row,
@@ -132,7 +155,6 @@ fn write_region(
 }
 
 fn main() -> Result<(), Error> {
-
     let matches = get_cli();
 
     let input_file = match matches.value_of("input_file") {
@@ -141,10 +163,10 @@ fn main() -> Result<(), Error> {
             if filepath.exists() {
                 filepath
             } else {
-                return Err(format_err!("Input file does not exist"))
+                return Err(format_err!("Input file does not exist"));
             }
         }
-        None => unreachable!()
+        None => unreachable!(),
     };
 
     let out_dir = match matches.value_of("out_root_dir") {
@@ -155,7 +177,7 @@ fn main() -> Result<(), Error> {
             }
             dirpath
         }
-        None => unreachable!()
+        None => unreachable!(),
     };
 
     let source_row = value_t!(matches.value_of("source_row"), u32)?;
@@ -179,8 +201,14 @@ fn main() -> Result<(), Error> {
     println!("Num levels: {}", os.get_level_count()?);
     println!("Dimensions at level 0: {:?}", os.get_level0_dimensions()?);
     println!("Dimensions at level 0: {:?}", os.get_level_dimensions(0)?);
-    println!("Downsample factor at level 0: {}", os.get_level_downsample(0)?);
-    println!("Best level for downsampling factor 4.5: {}", os.get_best_level_for_downsample(4.5)?);
+    println!(
+        "Downsample factor at level 0: {}",
+        os.get_level_downsample(0)?
+    );
+    println!(
+        "Best level for downsampling factor 4.5: {}",
+        os.get_best_level_for_downsample(4.5)?
+    );
 
     if matches.is_present("print_properties") {
         println!("Properties from dict");
@@ -200,9 +228,7 @@ fn main() -> Result<(), Error> {
         target_height,
         target_width,
         zoom_factor,
-        )?;
-
-
+    )?;
 
     Ok(())
 }